Moved pipe operation classes to extra package.
