+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/dsl/pipeops/PipeOperations.scala
@@ -1,4 +1,6 @@
-package io.shiftleft.queryprimitives.dsl
+package io.shiftleft.queryprimitives.dsl.pipeops
+
+import io.shiftleft.queryprimitives.dsl.RealPipe
 
 import scala.collection.GenTraversableOnce
