+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/steps/types/structure/MethodMethods.scala
@@ -3,10 +3,10 @@ package io.shiftleft.queryprimitives.steps.types.structure
 import gremlin.scala.Vertex
 import io.shiftleft.codepropertygraph.generated.nodes.NodeVisitor
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes, nodes}
-import io.shiftleft.queryprimitives.dsl.{PipeOperations, RealPipe}
+import io.shiftleft.queryprimitives.dsl.RealPipe
+import io.shiftleft.queryprimitives.dsl.pipeops.PipeOperations
 import io.shiftleft.queryprimitives.steps.ICallResolver
 import org.apache.tinkerpop.gremlin.structure.Direction
-
 
 import scala.collection.JavaConverters._
