+++ query-primitives/src/test/scala/io/shiftleft/queryprimitives/dsl/DslTests.scala
@@ -2,6 +2,7 @@ package io.shiftleft.queryprimitives.dsl
 
 import io.shiftleft.queryprimitives.dsl.Implicits._
 import io.shiftleft.queryprimitives.dsl.ShallowPipe.ShallowPipe
+import io.shiftleft.queryprimitives.dsl.pipeops.PipeOperations
 import org.scalatest.{Matchers, WordSpec}
 
 case class A()
